pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-video-pipeline"
version = "1.0.0"
description = "Automated AI video production pipeline with multi-platform publishing"
requires-python = ">=3.10"
dependencies = [
    "replicate>=0.25.0",
    "httpx>=0.27.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "instagrapi>=2.1.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "python-telegram-bot>=21.0",
    "vk-api>=11.9.0",
    "TikTokApi>=6.3.0",
    "moviepy>=1.0.3",
    "edge-tts>=6.1.0",
    "srt>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3.0",
]

[project.scripts]
videopipe = "pipeline.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.setuptools.packages.find]
include = ["pipeline*"]
